init work datahub separated plugin
