+++ .gitignore
@@ -90,5 +90,5 @@ web/src/main/webapp/doc/fr
 web/src/main/webapp/WEB-INF/data/data/resources/schemapublication
 
 # geonetwork-ui git project
-web/src/main/geonetwork-ui/
-web/node/
+plugins/datahub-integration/src/main/geonetwork-ui/
+plugins/datahub-integration/node/
