+++ plugins/datahub-integration/src/main/webResources/WEB-INF/config-security-mapping-datahub.xml
@@ -1,0 +1,45 @@
+<!--
+  ~ Copyright (C) 2001-2017 Food and Agriculture Organization of the
+  ~ United Nations (FAO-UN), United Nations World Food Programme (WFP)
+  ~ and United Nations Environment Programme (UNEP)
+  ~
+  ~ This program is free software; you can redistribute it and/or modify
+  ~ it under the terms of the GNU General Public License as published by
+  ~ the Free Software Foundation; either version 2 of the License, or (at
+  ~ your option) any later version.
+  ~
+  ~ This program is distributed in the hope that it will be useful, but
+  ~ WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+  ~ General Public License for more details.
+  ~
+  ~ You should have received a copy of the GNU General Public License
+  ~ along with this program; if not, write to the Free Software
+  ~ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA
+  ~
+  ~ Contact: Jeroen Ticheler - FAO - Viale delle Terme di Caracalla 2,
+  ~ Rome - Italy. email: geonetwork@osgeo.org
+  -->
+
+<beans
+  xsi:schemaLocation="http://www.springframework.org/schema/beans
+          http://www.springframework.org/schema/beans/spring-beans.xsd
+          http://www.springframework.org/schema/security
+          http://www.springframework.org/schema/security/spring-security.xsd"
+  xmlns:sec="http://www.springframework.org/schema/security"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns="http://www.springframework.org/schema/beans">
+  <bean class="org.springframework.security.web.access.intercept.FilterSecurityInterceptor"
+        id="datahubFilterInterceptor">
+    <property name="authenticationManager" ref="authenticationManager"></property>
+    <property name="accessDecisionManager" ref="accessDecisionManager"></property>
+    <property name="securityMetadataSource">
+      <sec:filter-security-metadata-source use-expressions="true" request-matcher="regex">
+        <sec:intercept-url
+          pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/datahub.*"
+          access="permitAll"/>
+        <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/datahub.*" access="permitAll"/>
+      </sec:filter-security-metadata-source>
+    </property>
+  </bean>
+</beans>
