+++ web/src/main/webapp/WEB-INF/config-security/config-security-mapping.xml
@@ -89,9 +89,6 @@
           pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/admin.config.csw.customelementset.save!?.*"
           access="hasAuthority('Administrator')"/>
 
-        <sec:intercept-url
-          pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/datahub.*"
-          access="permitAll"/>
         <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/admin.harvester.list!?.*"
                            access="hasAuthority('UserAdmin')"/>
         <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/admin.harvester.add!?.*"
@@ -257,9 +254,6 @@
         <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/generateSLD!?.*"
                            access="permitAll"/>
         <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/[a-z]{2,3}/getSLD/.*" access="permitAll"/>
-
-        <!-- Datahub endpoint -->
-        <sec:intercept-url pattern="/[a-zA-Z0-9_\-]+/datahub.*" access="permitAll"/>
 
         <sec:intercept-url pattern="/.*" access="denyAll"/>
